+++ mall-admin/src/main/java/cn/lzl/service/impl/PmsBrandServiceImpl.java
@@ -1,7 +1,5 @@
 package cn.lzl.service.impl;
-
-
-
+import cn.lzl.common.utils.PinyinHelperUtil;
 import cn.lzl.mapper.PmsBrandMapper;
 import cn.lzl.model.PmsBrand;
 import cn.lzl.model.PmsBrandExample;
@@ -10,9 +8,6 @@ import com.github.pagehelper.PageHelper;
 import org.springframework.stereotype.Service;
 import javax.annotation.Resource;
 import java.util.List;
-import net.sourceforge.pinyin4j.PinyinHelper;
-import net.sourceforge.pinyin4j.format.HanyuPinyinOutputFormat;
-import net.sourceforge.pinyin4j.format.HanyuPinyinToneType;
 @Service
 public class PmsBrandServiceImpl implements PmsBrandService {
 
@@ -23,31 +18,11 @@ public class PmsBrandServiceImpl implements PmsBrandService {
         return brandMapper.selectByExample(new PmsBrandExample());
     }
 
-    /**
-     * 得到中文首字母,例如"专科"得到zk返回
-     *
-     * @param str
-     *            中文字符串
-     * @return
-     */
-    public static String getPinYinHeadChar(String str) {
-        StringBuilder sb = new StringBuilder();
-        for (int i = 0; i < str.length(); i++) {
-            char word = str.charAt(i);
-            String[] pinyinArray = PinyinHelper.toHanyuPinyinStringArray(word);
-            if (pinyinArray != null) {
-                sb.append(pinyinArray[0].charAt(0));
-            } else {
-                sb.append(word);
-            }
-        }
-        return sb.toString().toUpperCase();
-    }
 
     @Override
     public int createBrand(PmsBrand brand) {
         if(null != brand.getFirstLetter() || ! "".equals(brand.getFirstLetter())){
-            String pinYinHeadChar = getPinYinHeadChar(brand.getFirstLetter());
+            String pinYinHeadChar = PinyinHelperUtil.getPinYinHeadChar(brand.getFirstLetter());
         }
         if(null == brand.getId() || "".equals(brand.getId())){
             return brandMapper.insertSelective(brand);
